requirements.txt
# Core AI and Agent Framework
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
langchain-community>=0.1.0

# CLI and User Interface
typer[all]>=0.9.0
rich>=13.0.0
click>=8.1.0

# Async and Utilities
asyncio-compat>=0.1.0
aiofiles>=23.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Data and JSON handling
json-schema>=4.0.0
jsonschema>=4.0.0

# Date and time utilities
python-dateutil>=2.8.0

# HTTP and networking (for tool simulations)
httpx>=0.25.0
requests>=2.31.0
